clean up main source code
