+++ src/paralleltraversal.cpp
@@ -36,78 +36,6 @@
 #define EXP 2.71828182845904523536
 
 
-/*! @brief The universal Levenshtein automaton for d=1.
-
-    The maximum length of a characteristic bitvector for d=1 is 2d+2=4.
-    The number of states of the this automaton is 15, these are numbered
-    as 0-14 and correspond to the following symbolic setup:\n
-    <table>
-     <tr><th> Table index </th><th> Symbolic state </th><th> Significance </th></tr>
-     <tr><td> 0 </td><td> {I^0} </td><td> non-accepting state </td></tr>
-     <tr><td> 1 </td><td> {(I-1)^1} </td><td> non-accepting state </td></tr>   
-     <tr><td> 2 </td><td> {I^1} </td><td> non-accepting state </td></tr>  
-     <tr><td> 3 </td><td> {(I-1)^1, I^1}} </td><td> non-accepting state </td></tr>  
-     <tr><td> 4 </td><td> {(I+1)^1}  </td><td> non-accepting state </td></tr> 
-     <tr><td> 5 </td><td> {(I-1)^1, (I+1)^1} </td><td> non-accepting state </td></tr> 
-     <tr><td> 6 </td><td> {I^1, (I+1)^1}  </td><td> non-accepting state </td></tr> 
-     <tr><td> 7 </td><td> {(I-1)^1, I^1, (I+1)^1}  </td><td> non-accepting state </td></tr> 
-     <tr><td> 8 </td><td> {(M-1)^0} </td><td> accepting state </td></tr> 
-     <tr><td> 9 </td><td> {(M-1)^0} </td><td> accepting state </td></tr> 
-     <tr><td> 10 </td><td> {M^1} </td><td> accepting state </td></tr> 
-     <tr><td> 11 </td><td> {(M-2)^1, M^1} </td><td> accepting state </td></tr>
-     <tr><td> 12 </td><td> {(M-1)^1, M^1} </td><td> accepting state </td></tr>
-     <tr><td> 13 </td><td> {(M-2)^1, (M-1)^1, M^1} </td><td> accepting state </td></tr>
-     <tr><td> 14 </td><td> NULL </td><td> failure state </td></tr>
-    </table>
- */
-uint32_t table[4][16][14] = {
-    {{3, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14},
-    {3, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14},
-    {7, 14, 14, 14, 4,  4,  4,  4,  14, 14, 14, 14, 14, 14},
-    {7, 14, 14, 14, 4,  4,  4,  4,  14, 14, 14, 14, 14, 14},
-    {0, 14, 2,  2,  14, 14, 2,  2,  14, 14, 14, 14, 14, 14},
-    {0, 14, 2,  2,  14, 14, 2,  2,  14, 14, 14, 14, 14, 14},
-    {0, 14, 2,  2,  4,  4,  6,  6,  14, 14, 14, 14, 14, 14},
-    {0, 14, 2,  2,  4,  4,  6,  6,  14, 14, 14, 14, 14, 14},
-    {3, 1,  14, 1, 14,  1, 14,  1,  14, 14, 14, 14, 14, 14},
-    {3, 1,  14, 1, 14,  1, 14,  1,  14, 14, 14, 14, 14, 14},
-    {7, 1,  14, 1, 4,   5, 4,   5,  14, 14, 14, 14, 14, 14},
-    {7, 1,  14, 1, 4,   5, 4,   5,  14, 14, 14, 14, 14, 14},
-    {0, 1,  2,  3, 14,  1, 2,   3,  14, 14, 14, 14, 14, 14},
-    {0, 1,  2,  3, 14,  1, 2,   3,  14, 14, 14, 14, 14, 14},
-    {0, 1,  2,  3, 4,   5, 6,   7,  14, 14, 14, 14, 14, 14},
-    {0, 1,  2,  3, 4,   5, 6,   7,  14, 14, 14, 14, 14, 14}},
-    {{3, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14},
-    {13, 14, 14, 14, 10, 10, 10, 10, 14, 14, 14, 14, 14, 14},
-    {8, 14, 2,  2,  14, 14, 2,  2,  14, 14, 14, 14, 14, 14},
-    {8, 14, 2,  2,  10, 10, 12, 12, 14, 14, 14, 14, 14, 14},
-    {3, 1, 14,  1,  14, 1,  14, 1,  14, 14, 14, 14, 14, 14},
-    {13, 1, 14, 1,  10, 11, 10, 11, 14, 14, 14, 14, 14, 14},
-    {8,  1,  2,  3,  14, 1,  2,  3, 14, 14, 14, 14, 14, 14},
-    {8, 1, 2, 3, 10, 11, 12, 13, 14, 14, 14, 14, 14, 14}},
-    {{12, 14, 14, 14, 14, 14, 14, 14, 12, 14, 14, 14, 14, 14},
-    {9, 14, 10, 10, 14, 14, 10, 10, 9, 14, 14, 14, 10, 10},
-    {12, 1, 14, 1, 14, 1, 14, 1, 12, 14, 14, 1, 14, 1},
-    {9, 1, 10, 12, 14, 1, 10, 12, 9, 14, 14, 1, 10, 12}},
-    {{10, 14, 14, 14, 14, 14, 14, 14, 14, 10, 14, 14, 14, 14},
-    {10, 10, 14, 10, 14, 10, 14, 10, 14, 10, 14, 14, 10, 14}}};
-
-/*! @brief Map nucleotides to integers.
-
-    Ambiguous letters map to 4. 
-    {A/a,C/c,G/g,T/t,U/u} = {0,1,2,3,3} respectively. 
-*/
-char nt_table[128] = {
-    4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,
-    4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,
-    4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,
-    4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,
-    4, 0, 4, 1,  4, 4, 4, 2,  4, 4, 4, 4,  4, 4, 4, 4,
-    4, 4, 4, 4,  3, 3, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4,
-    4, 0, 4, 1,  4, 4, 4, 2,  4, 4, 4, 4,  4, 4, 4, 4,
-    4, 4, 4, 4,  3, 3, 4, 4,  4, 4, 4, 4,  4, 4, 4, 4
-};
-
 /*! @brief Return complement of a nucleotide in
     integer format.
 
@@ -215,971 +143,6 @@ void format_rev(char* start_read,char* end_read,char* myread,char filesig)
 }
 
 
-/*! @fn traversetrie_align() */
-inline void
-traversetrie_align ( NodeElement *trie_t,
-                    uint32_t lev_t,
-                    unsigned char depth,
-                    MYBITSET *win_k1_ptr,
-                    MYBITSET *win_k1_full,
-                    bool &accept_zero_kmer,
-                    vector< id_win > &id_hits,
-                    uint32_t readn,
-                    uint32_t win_num,
-                    uint32_t partialwin
-                    )
-{
-  uint16_t lev_t_trie_pivot = lev_t;
-  unsigned char value = 0;
-    
-  // traverse through the node elements in a trie node
-  for ( uint32_t node_element = 0; node_element < 4; node_element++ )
-  {
-    value = trie_t->flag;
-        
-    // this node element is empty, go to next node element in trie node
-    if ( value == 0 )
-    {
-      lev_t = lev_t_trie_pivot;
-      trie_t++;
-    }
-        
-    // this node element points to a trie node or a bucket, continue traversing
-    else
-    {
-      if ( depth < partialwin-2 )
-      {
-        // send bv to LEV(1)
-        lev_t = table[0][(int)*(win_k1_ptr + (depth<<2) + node_element)][(int)(lev_t)];
-      }
-      else
-      {
-        lev_t = table[3-partialwin+depth][(int)(*(win_k1_full+node_element) & ((2<<(partialwin-depth))-1))][(int)(lev_t)];
-      }
-            
-            
-      // LEV(1) is in a null state, go to next node element
-      if ( lev_t == 14 )
-      {
-        lev_t = lev_t_trie_pivot;
-        trie_t++;
-      }
-      // LEV(1) is not in a null state, continue parallel traversal
-      else
-      {
-        // (1) the node element holds a pointer to another trie node
-        if ( value == 1 )
-        {
-          traversetrie_align ( trie_t->whichnode.trie,
-                                        lev_t,
-                                        ++depth,
-                                        win_k1_ptr,
-                                        win_k1_full,
-                                        accept_zero_kmer,
-                                        id_hits,
-                                        readn,
-                                        win_num,
-                                        partialwin
-                                        );
-                    
-          // go to next window on the read (0-error match found)
-          if ( accept_zero_kmer ) return ;
-                    
-          --depth;
-                    
-          // go to the next trie node element
-          lev_t = lev_t_trie_pivot;
-          trie_t++;
-        }//~ if (child node a trie node)                
-        // (2) the node element points to a bucket
-        else
-        {
-          // this pivot remembers the target levenshtein state of the terminal trie node, from
-          //  which exists a bucket; every element in the bucket takes this lev_t state as an
-          //  initial input state
-          uint32_t lev_t_bucket_pivot = lev_t;
-                    
-          // number of characters per entry
-          uint32_t s = partialwin-depth;
-                    
-          unsigned char* start_bucket = (unsigned char*)trie_t->whichnode.bucket;
-          if ( start_bucket == NULL )
-          {
-            fprintf(stderr, "  ERROR: pointer start_bucket == NULL (paralleltraversal.cpp)\n");
-            exit(EXIT_FAILURE);
-          }
-          unsigned char* end_bucket = start_bucket + trie_t->size;
-          if ( end_bucket == NULL )
-          {
-            fprintf(stderr, "  ERROR: pointer end_bucket == NULL (paralleltraversal.cpp)\n");
-            exit(EXIT_FAILURE);
-          }
-                    
-          // traverse the bucket
-          while ( start_bucket != end_bucket )
-          {
-            uint32_t depth_b = depth;
-            lev_t = lev_t_bucket_pivot;
-            bool local_accept_kmer = false;
-            uint32_t entry_str = *((uint32_t*)start_bucket);
-                        
-            // for each nt in the string
-            for ( uint32_t j = 0; j < s; j++ )
-            {
-              uint32_t nt = entry_str&3;
-                            
-              depth_b++;
-                            
-              // get bitvector for letter
-              if ( depth_b < partialwin-2 )
-              {
-                // send bv to LEV(_k)
-                lev_t = table[0][(int)*(win_k1_ptr + (depth_b<<2) + nt)][(int)(lev_t)];
-              }
-              else
-              {
-                lev_t = table[3-partialwin+depth_b][(int)(*(win_k1_full + nt) & ((2<<(partialwin-depth_b))-1) )][(int)(lev_t)];
-              }
-                            
-              // if the target lev_t state is a failure state, go to the next bucket element (tail)
-              if ( lev_t == 14 ) break;
-                            
-              // approaching end of tail
-              if ( depth_b >= partialwin-2 )
-              {
-                // 1-error match
-                if ( lev_t >= 8 )
-                {
-                  local_accept_kmer = true;
-                }
-                // 0-error match
-                if ( depth_b == partialwin-1 )
-                {
-                  if ( lev_t == 9 )
-                  {
-                    accept_zero_kmer = true;
-                                        
-                    // turn off heuristic to stop search after finding 0-error match
-                    if ( full_search_gv ) accept_zero_kmer = false;
-                  }
-                }
-              }//~last 3 characters in entry
-                            
-              if ( local_accept_kmer )
-              {
-                id_win entry = {0,0};
-                entry.id = *((uint32_t*)start_bucket + 1);
-                entry.win = win_num;
-                                
-                // empty id_hits array, add 0-error id and exit
-                if ( accept_zero_kmer )
-                {
-                  id_hits.clear();
-                  id_hits.push_back(entry);
-                                    
-                  return;
-                }
-                                
-                // exact match not found, do not include duplicates of 1-error match (for the same window on read)
-                if ( !id_hits.empty() )
-                {
-                  bool found = false;
-                  for ( int f = 0; f < id_hits.size(); f++ )
-                  {
-                    if ( id_hits[f].id == entry.id )
-                    {
-                      found = true;
-                      break;
-                    }
-                  }
-                  if ( found ) break;
-                }
-                                
-                id_hits.push_back(entry);
-                                
-              }                         
-              entry_str>>=2;
-            }//~for each 2 bits
-                        
-            // next entry
-            start_bucket+=ENTRYSIZE;
-          }//~for each entry
-                    
-          lev_t = lev_t_trie_pivot;
-          trie_t++;
-                    
-        }//~else the node element points to a bucket
-      }//~else LEV(1) is not in a null state, continue parallel traversal
-    }//~else this node element points to a trie node or a bucket, continue traversing
-  }//~for all trie nodes
-    
-  return ;
-}//~traversetrie_align()
-
-
-
-
-/*
- *
- * @function check files: load the @SQ lines for SAM output & input all header information to SAM output file
- * @return void
- * @version 2.0 February 4, 2014
- *
- *******************************************************************/
-void preprocess_data(vector< pair<string,string> >& myfiles,
-                     char** argv,
-                     int argc,
-                     bool yes_SQ,
-                     char* acceptedstrings_sam,
-                     int32_t _match,
-                     int32_t _mismatch,
-                     int32_t _gap_open,
-                     int32_t _gap_extension,
-                     vector<vector<uint32_t> >& skiplengths,
-                     vector<uint16_t>& num_index_parts,
-                     vector<vector<index_parts_stats> >& index_parts_stats_vec,
-                     vector<uint64_t>& full_ref,
-                     vector<uint64_t>& full_read,
-                     vector<uint32_t>& lnwin,
-                     vector<uint32_t>& partialwin,
-                     vector<uint32_t>& minimal_score,
-                     uint32_t number_total_read,
-                     vector<pair<double, double> >& gumbel,
-                     vector<uint32_t>& numbvs,
-                     vector<uint32_t>& numseq)
-{
-  ofstream acceptedsam;
-  
-  if ( samout_gv )
-  {
-    acceptedsam.open (acceptedstrings_sam);
-    if (!acceptedsam.good())
-    {
-      fprintf(stderr,"  %sERROR%s: could not open SAM output file for writing.\n","\033[0;31m","\033[0m");
-      exit(EXIT_FAILURE);
-    }
-    // @HD header
-    else acceptedsam << "@HD\tVN:1.0\tSO:unsorted\n";
-  }
-    
-  // loop through the .seqs index files for each database
-  for ( uint16_t index_num = 0; index_num < (uint16_t)myfiles.size(); index_num++ )
-  {
-      ifstream stats( (char*)(myfiles[index_num].second + ".stats").c_str(), ios::in | ios::binary );
-      if ( !stats.good() )
-      {
-        fprintf(stderr,"\n  %sERROR%s: The index '%s' does not exist.\n","\033[0;31m","\033[0m",(char*)(myfiles[index_num].second + ".stats").c_str());
-        fprintf(stderr,"  Make sure you have constructed your index using the command `indexdb'. See `indexdb -h' for help.\n\n");
-        exit(EXIT_FAILURE);
-      }
-      
-      // read the file size for file used to build the index
-      size_t filesize = 0;
-      stats.read(reinterpret_cast<char*>(&filesize), sizeof(size_t));
-      
-      // read the fasta file name used to build the index
-      uint32_t fastafile_len = 0;
-      stats.read(reinterpret_cast<char*>(&fastafile_len), sizeof(uint32_t));
-      
-      char fastafile_name[2000];
-      stats.read(reinterpret_cast<char*>(fastafile_name), sizeof(char)*fastafile_len);
-      
-      // compute reference database file size for this index
-      FILE *fastafile = fopen ((char*)(myfiles[index_num].first).c_str(),"r");
-      if ( fastafile == NULL )
-      {
-        fprintf(stderr,"    %sERROR%s: could not open FASTA reference file: %s .\n","\033[0;31m","\033[0m",(char*)(myfiles[index_num].first).c_str());
-        exit(EXIT_FAILURE);
-      }
-      fseek(fastafile,0L,SEEK_END);
-      size_t sz = ftell(fastafile);
-      fclose(fastafile);
-        
-      if ( sz != filesize )
-      {
-        fprintf(stderr,"    %sERROR%s: Based on file size, the FASTA file (%s) passed to --ref <FASTA file, index name>\n","\033[0;31m","\033[0m",(char*)(myfiles[index_num].first).c_str());
-        fprintf(stderr,"    does not appear to be the same FASTA file (%s) used to build the index %s.\n",fastafile_name,(char*)(myfiles[index_num].second).c_str());
-        fprintf(stderr,"    Check your --ref list of files and corresponding indexes.\n\n");
-        exit(EXIT_FAILURE);
-      }
-      
-      // A,C,G,T background frequencies to compute the Gumbel parameters lambda and K
-      double background_freq_gv[4] = {0};
-      
-      // A/C/G/T distribution frequencies
-      stats.read(reinterpret_cast<char*>(&background_freq_gv), sizeof(double)*4);
-      
-      // total length of sequences in the complete database
-      stats.read(reinterpret_cast<char*>(&full_ref[index_num]), sizeof(uint64_t));
-      
-      // sliding window length lnwin & initialize
-      stats.read(reinterpret_cast<char*>(&lnwin[index_num]), sizeof(uint32_t));
-      
-      // total number of reference sequences in one complete reference database
-      stats.read(reinterpret_cast<char*>(&numseq[index_num]), sizeof(uint32_t));
-      
-      partialwin[index_num] = lnwin[index_num]/2;
-        
-      // number of bitvectors at depth > 0 in [w_1] reverse or [w_2] forward
-      numbvs[index_num] = 4*(partialwin[index_num]-3);
-      
-      // set the window shift for different seed lengths (if not set by user, or one of the lengths is <= 0)
-      if ( (skiplengths[index_num][0] == 0) || (skiplengths[index_num][1] == 0) || (skiplengths[index_num][2] == 0) )
-      {
-        skiplengths[index_num][0] = lnwin[index_num];
-        skiplengths[index_num][1] = partialwin[index_num];
-        skiplengths[index_num][2] = 3;
-      }
-      
-      // number of index parts
-      stats.read(reinterpret_cast<char*>(&num_index_parts[index_num]), sizeof(uint16_t));
-      vector<index_parts_stats> hold;
-      
-      // information on the location and size of sequences used to build each index part
-      for ( uint16_t j = 0; j < num_index_parts[index_num]; j++ )
-      {
-        index_parts_stats stats_hold;
-        stats.read(reinterpret_cast<char*>(&stats_hold), sizeof(index_parts_stats));
-        hold.push_back(stats_hold);
-      }
-        
-      index_parts_stats_vec.push_back(hold);
-      
-      // compute Gumbel parameters
-      long int rand_ = 182345345;
-      //string randout_= "./alp/random_param.txt";
-      string randout_= "";
-      
-      int gapopen_ = _gap_open;
-      int gapopen1_ = _gap_open;
-      int gapopen2_ = _gap_open;
-      
-      int gapextend_ = _gap_extension;
-      int gapextend1_ = _gap_extension;
-      int gapextend2_ = _gap_extension;
-      
-      int match = _match;
-      int mismatch = _mismatch;
-      double A_ = background_freq_gv[0];
-      double C_ = background_freq_gv[1];
-      double G_ = background_freq_gv[2];
-      double T_ = background_freq_gv[3];
-      
-      string scoremat_file_name_ ="";
-      string freqs1_file_name_ ="";
-      string freqs2_file_name_ ="";
-      double max_time_=1;
-      double max_mem_=500;
-      double eps_lambda_gv_=0.001;
-      double eps_K_gv_=0.005;
-      string gumbelparout_file_name_ ="";
-      bool gapped_ = true;
-      bool insertions_after_deletions_=false;
-      
-      Sls::set_of_parameters gumbel_params;
-        
-      CGumbelParamsCalc::Params_Run2(
-                                     rand_,//randomization number
-                                     randout_,//if true, then the program outputs complete randomization information into a file
-                                     
-                                     gapopen_,//gap opening penalty
-                                     gapopen1_,//gap opening penalty for a gap in the sequence #1
-                                     gapopen2_,//gap opening penalty for a gap in the sequence #2
-                                     
-                                     gapextend_,//gap extension penalty
-                                     gapextend1_,//gap extension penalty for a gap in the sequence #1
-                                     gapextend2_,//gap extension penalty for a gap in the sequence #2
-                                     
-                                     scoremat_file_name_,//scoring matrix file name
-                                     freqs1_file_name_,//probabilities1 file name
-                                     freqs2_file_name_,//probabilities1 file name
-                                     max_time_,//maximum allowed calculation time in seconds
-                                     max_mem_,//maximum allowed memory usage in MB
-                                     eps_lambda_gv_,//relative error for lambda_gv calculation
-                                     eps_K_gv_,//relative error for K_gv calculation
-                                     gumbelparout_file_name_,
-                                     gapped_,
-                                     insertions_after_deletions_,//if true, then insertions after deletions are allowed
-                                     gumbel_params,
-                                     match,//NEW - SW score for a match,
-                                     mismatch,//NEW - SW score for a mismatch,
-                                     A_,//NEW - background frequency for A
-                                     C_,//NEW - background frequency for C
-                                     G_,//NEW - background frequency for G
-                                     T_,//NEW- background frequency for T
-                                     (gumbel[index_num].first), //lambda
-                                     (gumbel[index_num].second) // K
-                                     );
-        
-        
-      // Shannon's entropy for reference sequence nucleotide distribution
-      double entropy_H_gv = -(background_freq_gv[0]*(log(background_freq_gv[0])/log(2)) +
-                              background_freq_gv[1]*(log(background_freq_gv[1])/log(2)) +
-                              background_freq_gv[2]*(log(background_freq_gv[2])/log(2)) +
-                              background_freq_gv[3]*(log(background_freq_gv[3])/log(2)));
-      
-      // Length correction for Smith-Waterman alignment score
-      uint64_t expect_L = log((gumbel[index_num].second)*full_read[index_num]*full_ref[index_num])/entropy_H_gv;
-      
-      // correct the reads & databases sizes for e-value calculation
-      if ( full_ref[index_num] > (expect_L*numseq[index_num]) ) full_ref[index_num]-=(expect_L*numseq[index_num]);
-      full_read[index_num]-=(expect_L*number_total_read);
-      
-      // minimum score required to reach E-value
-      minimal_score[index_num] = (log(evalue/((double)(gumbel[index_num].second)*full_ref[index_num]*full_read[index_num])))/-(gumbel[index_num].first);
-        
-      // SAM @SQ data
-      if ( samout_gv )
-      {
-        // number of nucleotide sequences in the reference file
-        uint32_t num_sq = 0;
-        stats.read(reinterpret_cast<char*>(&num_sq), sizeof(uint32_t));
-        
-        // loop through each @SQ
-        for ( uint32_t j = 0; j < num_sq; j++ )
-        {
-          // length of the sequence id
-          uint32_t len_id = 0;
-          stats.read(reinterpret_cast<char*>(&len_id), sizeof(uint32_t));
-          
-          // the sequence id string
-          char s[len_id+1];
-          memset(s,0,len_id+1);
-          stats.read(reinterpret_cast<char*>(&s), sizeof(char)*len_id);
-          
-          // the length of the sequence itself
-          uint32_t len_seq = 0;
-          stats.read(reinterpret_cast<char*>(&len_seq), sizeof(uint32_t));
-          
-          // @SQ header
-          if ( yes_SQ ) acceptedsam << "@SQ\tSN:" << s << "\tLN:" << len_seq << "\n";
-        }
-      }
-        
-      stats.close();
-    }//~for all indexes
-    
-    if ( samout_gv )
-    {
-      // @PG to sam file
-      acceptedsam << "@PG\tID:sortmerna\tVN:1.0\tCL:";
-      for ( int j = 0; j < argc; j++ ) acceptedsam << argv[j] << " ";
-      acceptedsam << endl;
-      
-      acceptedsam.close();
-    }//~if samout_gv
-}//~preprocess_data()
-
-
-
-
-
-/*
- *
- * @function load_ref: load fasta reference sequences into memory
- * for SW alignment
- * @return void
- * @version 1.0 June 12, 2013
- *
- *******************************************************************/
-void
-load_ref(char* ptr_dbfile,
-         char* buffer,
-         char** reference_seq,
-         uint32_t* reference_seq_len,
-         uint64_t seq_part_size,
-         uint32_t numseq_part,
-         uint64_t start_part,
-         bool load_for_search)
-{
-  FILE *fp = fopen(ptr_dbfile,"r");
-  if ( fp == NULL )
-  {
-    fprintf(stderr,"  %sERROR%s: could not open file %s\n",ptr_dbfile,"\033[0;31m","\033[0m");
-    exit(EXIT_FAILURE);
-  }
-    
-  // set the file pointer to the first sequence added to the index for this index file section
-  if ( fseek(fp,start_part,SEEK_SET) != 0 )
-    {
-      fprintf(stderr,"  %sERROR%s: could not locate the sequences used to construct the index (paralleltraversal.cpp).\n","\033[0;31m","\033[0m");
-      fprintf(stderr,"  Check that your --ref <FASTA file, index name> correspond correctly for the FASTA file: %s.\n",ptr_dbfile);
-    }
-    
-  // load references sequences into memory, skipping the new lines & spaces in the fasta format
-  uint32_t num_seq_read = 0;
-  char *s = buffer;
-    
-  int i = 0;
-  int j = 0;
-  char c = fgetc(fp);
-
-  if ( load_for_search )
-  {
-    do
-    {
-      // the tag
-      reference_seq[i++] = s;
-      while ( c != '\n' )
-      {
-        *s++ = c;
-        c = fgetc(fp);
-      }
-      // new line
-      *s++ = c;
-      
-      if ( *s == '\n' )
-      {
-        fprintf(stderr,"  %sERROR%s: your reference sequences are not in FASTA format "
-                       "(there is an extra new line).","\033[0;31m","\033[0m");
-        exit(EXIT_FAILURE);
-      }
-      
-      // the sequence
-      reference_seq[i++] = s;
-      c = fgetc(fp);
-      do
-      {
-        if ( c != '\n' && c != ' ' )
-        {
-          // keep record of ambiguous character for alignment
-          *s++ = nt_table[(int)c];
-          
-          // record the sequence length as we read it
-          reference_seq_len[j]++;
-        }
-        c = fgetc(fp);
-      } while ( (c != '>') && (c != EOF) );
-      
-      *s++ = '\n';
-      j++;
-      
-      num_seq_read++;
-        
-    } while ( (num_seq_read != numseq_part) && (c != EOF) );
-  }
-  else
-  {
-    do
-    {
-      // the tag
-      reference_seq[i++] = s;
-      while ( c != '\n' )
-      {
-        *s++ = c;
-        c = fgetc(fp);
-      }
-      
-      // new line
-      *s++ = c;
-      
-      // the sequence
-      reference_seq[i++] = s;
-      c = fgetc(fp);
-      do
-      {
-        if ( c != '\n' && c != ' ' )
-        {
-            // keep record of ambiguous character for alignment
-            *s++ = nt_table[(int)c];
-        }
-        c = fgetc(fp);
-      } while ( (c != '>') && (c != EOF) );
-      
-      *s++ = '\n';
-      j++;
-      
-      num_seq_read++;
-        
-    } while ( (num_seq_read != numseq_part) && (c != EOF) );
-  }
-    
-  fclose(fp);
-}//~load_ref
-
-
-#ifdef see_binary_output
-/*
- *
- * @function traversetrie: collect statistics on the mini-burst trie,
- * its size, number of trie nodes vs. buckets
- * @param NodeElement* trie_node
- * @return void
- * @version 1.0 Jan 14, 2013
- *
- *******************************************************************/
-void traversetrie_debug( NodeElement* trie_node, uint32_t depth, uint32_t &total_entries, string &kmer_keep, uint32_t partialwin )
-{
-  char get_char[4] = {'A','C','G','T'};
-    
-  // traverse through the node elements in a trie node
-  for ( int i = 0; i < 4; i++ )
-  {
-    unsigned char value = trie_node->flag;
-        
-    // the node element holds a pointer to another trie node
-    if ( value == 1 )
-    {
-      kmer_keep.push_back((char)get_char[i]); //TESTING
-      traversetrie_debug( trie_node->whichnode.trie, ++depth, total_entries, kmer_keep, partialwin );
-      kmer_keep.pop_back();
-      --depth;
-    }
-        
-    // the node element points to a bucket
-    else if ( value == 2 )
-    {
-      kmer_keep.push_back((char)get_char[i]); //TESTING
-      
-      unsigned char* start_bucket = (unsigned char*)trie_node->whichnode.bucket;
-      if ( start_bucket == NULL )
-      {
-        fprintf(stderr, "  ERROR: pointer start_bucket == NULL (paralleltraversal.cpp)\n");
-        exit(EXIT_FAILURE);
-      }
-      unsigned char* end_bucket = start_bucket + trie_node->size;
-      if ( end_bucket == NULL )
-      {
-        fprintf(stderr, "  ERROR: pointer end_bucket == NULL (paralleltraversal.cpp)\n");
-        exit(EXIT_FAILURE);
-      }
-      
-      cout << "size of bucket = " << trie_node->size << endl; //TESTING
-      cout << "end_bucket-start_bucket = " << (end_bucket-start_bucket) << endl; //TESTING
-            
-      // traverse the bucket
-      while ( start_bucket != end_bucket )
-      {
-        uint32_t entry_str = *((uint32_t*)start_bucket);
-        uint32_t s = partialwin-depth;
-        total_entries++;
-        
-        // for each nt in the string
-        for ( uint32_t j = 0; j < s; j++ )
-        {
-          kmer_keep.push_back((char)get_char[entry_str&3]); //TESTING
-          entry_str>>=2;
-        }
-        
-        cout << kmer_keep << endl; //TESTING
-        
-        for ( uint32_t j = 0; j < s; j++ ) kmer_keep.pop_back(); //TESTING
-        
-        // next entry
-        start_bucket+=ENTRYSIZE;
-      }//~for each entry
-      
-      kmer_keep.pop_back(); //TESTING
-            
-    }
-        
-    // the node element is empty, go to next node element
-    else if ( value == 0 )
-    {
-      ;
-    }
-        
-    trie_node++;
-        
-  }//~trie nodes
-    
-  return;
-    
-}//~taversetrie_debug()
-#endif
-
-
-
-
-/*
- *
- * @function load index: read from binary file the L/2-mer look-up
- * tables, the 19-mer position tables and the mini-burst tries
- * @return void
- * @version 1.0 Jan 16, 2013
- *
- *******************************************************************/
-void
-load_index( char* ptr_dbindex,
-           string part_str,
-           kmer*& lookup_tbl,
-           kmer_origin*& positions_tbl,
-           uint32_t& number_elements,
-           uint32_t lnwin)
-{
-    // STEP 1: load the kmer 'count' variables (dbname.kmer.dat)
-    string ptr_dbindex_str = ptr_dbindex;
-  ifstream inkmer( (char*)(ptr_dbindex_str + ".kmer_" + part_str + ".dat").c_str(), ios::in | ios::binary );
-    
-  if ( !inkmer.good() )
-  {
-    fprintf(stderr,"\n  ERROR: The index '%s' does not exist.\n", (char*)(ptr_dbindex_str + ".kmer_" + part_str + ".dat").c_str());
-    fprintf(stderr,"  Make sure you have constructed your index using the command `indexdb'. See `indexdb -h' for help.\n\n");
-    exit(EXIT_FAILURE);
-  }
-    
-  lookup_tbl = new kmer[(1<<lnwin)]();
-  if ( lookup_tbl == NULL )
-  {
-    fprintf(stderr,"\n  ERROR: failed to allocate memory for look-up table (paralleltraversal.cpp)\n\n");
-    exit(EXIT_FAILURE);
-  }
-    
-  uint32_t limit = 1<<lnwin;
-    
-  for ( uint32_t i = 0; i < limit; i++ ) inkmer.read(reinterpret_cast<char*>(&(lookup_tbl[i].count)), sizeof(uint32_t));
-    
-  inkmer.close();
-    
-  // STEP 2: load the burst tries ( bursttrief.dat, bursttrier.dat )
-  ifstream btrie( (char*)(ptr_dbindex_str + ".bursttrie_" + part_str + ".dat").c_str(), ios::in | ios::binary );
-    
-  if ( !btrie.good() )
-  {
-    fprintf(stderr,"\n  ERROR: The index '%s' does not exist.\n", (char*)(ptr_dbindex_str + ".bursttrie_" + part_str + ".dat").c_str());
-    fprintf(stderr,"  Make sure you have constructed your index using the command `indexdb'. See `indexdb -h' for help.\n\n");
-    exit(EXIT_FAILURE);
-  }
-    
-  // loop through all 9-mers
-  for ( uint32_t i = 0; i < (uint32_t)(1<<lnwin); i++ )
-  {
-    uint32_t sizeoftries[2] = {0};
-        
-#ifdef see_binary_output
-        cout << "9-mer = " << i; //TESTING
-#endif
-        
-    // ptr to block of memory for two mini-burst tries
-    char *dst = NULL;
-        
-    // the size of both mini-burst tries
-    for ( int j = 0; j < 2; j++ )
-    {
-      btrie.read(reinterpret_cast<char*>(&sizeoftries[j]), sizeof(uint32_t));
-            
-#ifdef see_binary_output
-            if ( j == 0 ) cout << "\tsizeoftrie f = " << sizeoftries[j]; //TESTING
-            else cout << "\tsizeoftrie r = " << sizeoftries[j]; //TESTING
-#endif
-            
-    }
-        
-#ifdef see_binary_output
-        cout << "\tlookup_tbl[i].count = " << lookup_tbl[i].count << endl; //TESTING
-#endif
-        
-    // allocate contiguous memory for both mini-burst tries if they exist
-    if ( lookup_tbl[i].count != 0 )
-    {
-      dst = new char[(sizeoftries[0]+sizeoftries[1])]();
-      if ( dst == NULL )
-      {
-        fprintf(stderr,"  ERROR: failed to allocate memory for mini-burst tries (paralleltraversal.cpp)\n");
-        exit(EXIT_FAILURE);
-      }
-            
-      // load 2 burst tries per 9-mer
-      for ( int j = 0; j < 2; j++ )
-      {
-        // mini-burst trie exists
-        if ( sizeoftries[j] != 0 )
-        {
-#ifdef see_binary_output
-          if ( j == 0 ) cout << "forward burst-trie \n"; //TESTING
-          if ( j == 1 ) cout << "reverse burst-trie \n"; //TESTING
-#endif            
-          // create a root trie node
-          NodeElement newnode[4];
-                    
-          // copy the root trie node into the beginning of burst trie array
-          memcpy( dst, &newnode[0], sizeof(NodeElement)*4);
-          memset( dst, 0, sizeof(NodeElement)*4);
-                    
-          if ( j == 0 ) lookup_tbl[i].trie_F = (NodeElement*)dst;
-          else lookup_tbl[i].trie_R = (NodeElement*)dst;
-                    
-          // queue to store the trie nodes as we create them
-          deque<NodeElement*> nodes;
-          nodes.push_back( (NodeElement*)dst );
-          ((NodeElement *&) dst)+=4;
-                    
-          // queue to store the flags of node elements given in the binary file
-          deque<char> flags;
-                    
-          // read the first trie node
-          for ( int i = 0; i < 4; i++ )
-          {
-            char tmp;
-            btrie.read(reinterpret_cast<char*>(&tmp), sizeof(char));
-            flags.push_back(tmp);
-#ifdef see_binary_output
-                        cout << " " << (int)tmp; //TESTING
-#endif
-          }
-                    
-          // build the mini-burst trie
-          while ( !nodes.empty() )
-          {
-            // ptr to traverse each trie node
-            NodeElement* node = nodes.front();
-                        
-            // trie node elements
-            for ( int i = 0; i < 4; i++ )
-            {
-              unsigned char flag = flags.front();
-                            
-              // what does the node element point to
-              switch( flag )
-              {
-                // set values to 0
-                case 0:
-                {
-                  node->flag = 0;
-                  node->size = 0;
-                  node->whichnode.trie = NULL;
-                }
-                  break;
-                // trie node
-                case 1:
-                {
-                  // read the trie node
-                  for ( int i = 0; i < 4; i++ )
-                  {
-                    char tmp;
-                    btrie.read(reinterpret_cast<char*>(&tmp), sizeof(char));
-#ifdef see_binary_output
-                    cout << " " << (int)tmp; //TESTING
-#endif
-                    flags.push_back(tmp);
-                  }
-                                    
-                  node->flag = 1;
-                  node->size = 0;
-                  NodeElement newnode[4];
-                  memcpy( (NodeElement*)dst, &newnode[0], sizeof(NodeElement)*4);
-                  nodes.push_back( (NodeElement*)dst );
-                  node->whichnode.trie = (NodeElement*)dst;
-                  ((NodeElement *&) dst)+=4;
-                                    
-                }
-                  break;
-                // bucket
-                case 2:
-                {
-                  uint32_t sizeofbucket = 0;
-                                    
-                  // read the bucket info
-                  btrie.read(reinterpret_cast<char*>(&sizeofbucket), sizeof(uint32_t));
-                                    
-#ifdef see_binary_output
-                  cout << "\tsizeofbucket = " << sizeofbucket; //TESTING
-#endif                      
-                  char* bucket = new char[sizeofbucket]();
-                  if ( bucket == NULL )
-                  {
-                    fprintf(stderr,"\n  %sERROR%s: failed to allocate memory for allocate bucket (paralleltraversal.cpp)\n","\033[0;31m","\033[0m");
-                    exit(EXIT_FAILURE);
-                  }
-                                    
-                  btrie.read(reinterpret_cast<char*>(bucket), sizeofbucket);
-                                    
-                  // copy the bucket into the burst trie array
-                  memcpy( (void*)dst, (void*)bucket, sizeofbucket);
-                                    
-                  delete [] bucket;
-                  bucket = NULL;
-                                    
-                  // assign pointers from trie node to the bucket
-                  node->flag = flag;
-                  node->whichnode.bucket = dst;
-                  node->size = sizeofbucket;
-                  dst = ((char *)dst)+sizeofbucket;        
-                }
-                  break;
-                // ?
-                default:
-                {
-                  fprintf(stderr, "\n  %sERROR%s: flag is set to %d (load_index)\n","\033[0;31m","\033[0m",flag);
-                  exit(EXIT_FAILURE);
-                }
-                  break;
-              }
-
-              flags.pop_front();
-              node++;               
-            }//~loop through 4 node elements in a trie node
-                        
-#ifdef see_binary_output
-            cout << "\n"; //TESTING
-#endif            
-            nodes.pop_front();
-                        
-          }//~while !nodes.empty()
-        }//~if mini-burst trie exists
-        else
-        {
-          if ( j == 0 ) lookup_tbl[i].trie_F = NULL;
-          else lookup_tbl[i].trie_R = NULL;
-        }
-      }//~for both mini-burst tries
-    }//~if ( sizeoftries != 0 )
-    else
-    {
-      lookup_tbl[i].trie_F = NULL;
-      lookup_tbl[i].trie_R = NULL;
-    }
-  }//~for all 9-mers in the look-up table
-    
-  btrie.close();
-    
-  // STEP 3: load the position reference tables (pos.dat)
-  ifstream inreff( (char*)(ptr_dbindex_str + ".pos_" + part_str + ".dat").c_str(), ios::in | ios::binary );
-    
-  if ( !inreff.good() )
-  {
-    fprintf(stderr,"\n  ERROR: The database name '%s' does not exist.\n\n", (char*)(ptr_dbindex_str + ".pos_" + part_str + ".dat").c_str());
-    exit(EXIT_FAILURE);
-  }
-    
-  uint32_t size = 0;
-    
-  inreff.read(reinterpret_cast<char*>(&number_elements), sizeof(uint32_t));
-    
-  positions_tbl = new kmer_origin[number_elements]();
-  if ( positions_tbl == NULL )
-  {
-    fprintf(stderr,"  ERROR: could not allocate memory for positions_tbl (main(), paralleltraversal.cpp)\n");
-    exit(EXIT_FAILURE);
-  }
-    
-  for ( uint32_t i = 0; i < number_elements; i++ )
-  {
-    /* the number of positions */
-    inreff.read(reinterpret_cast<char*>(&size), sizeof(uint32_t));
-    positions_tbl[i].size = size;
-        
-    /* the sequence seq_pos array */
-    positions_tbl[i].arr = new seq_pos[size]();
-        
-    if ( positions_tbl[i].arr == NULL )
-    {
-      fprintf(stderr, "  ERROR: could not allocate memory for positions_tbl (paralleltraversal.cpp)\n");
-      exit(EXIT_FAILURE);
-    }
-        
-    inreff.read(reinterpret_cast<char*>(positions_tbl[i].arr), sizeof(seq_pos)*size);
-  }
-    
-  inreff.close();
-    
-  return ;
-    
-}//~load_index()
-
-
-
-
 /*
  *
  * FUNCTION   : find_lis()
@@ -1238,19 +201,19 @@ void find_lis( deque<pair<uint32_t, uint32_t> > &a, vector<uint32_t> &b, uint32_
 /*! @fn paralleltraversal() */
 void
 paralleltraversal ( char* inputreads,
-                   char* ptr_filetype_ar,
-                   char* ptr_filetype_or,
-                   int32_t match,
-                   int32_t mismatch,
-                   int32_t gap_open,
-                   int32_t gap_extension,
-                   int32_t score_N,
-                   vector< vector<uint32_t> >& skiplengths,
-                   int argc,
-                   char **argv,
-                   bool yes_SQ,
-                   vector< pair<string,string> >& myfiles,
-                   bool exit_early)
+                    char* ptr_filetype_ar,
+                    char* ptr_filetype_or,
+                    int32_t match,
+                    int32_t mismatch,
+                    int32_t gap_open,
+                    int32_t gap_extension,
+                    int32_t score_N,
+                    vector< vector<uint32_t> >& skiplengths,
+                    int argc,
+                    char **argv,
+                    bool yes_SQ,
+                    vector< pair<string,string> >& myfiles,
+                    bool exit_early)
 {
   // the offset from the start of the reads file for mmap
   off_t offset_map = 0;
@@ -1638,7 +601,7 @@ paralleltraversal ( char* inputreads,
   vector<uint64_t> full_read(myfiles.size(),full_read_main);
   vector<uint32_t> lnwin(myfiles.size(),0);
   vector<uint32_t> partialwin(myfiles.size(),0);
-  // Minimal SW score in order to reach threshold E-value
+  // minimal SW score in order to reach threshold E-value
   vector<uint32_t> minimal_score(myfiles.size(),0);
   // array of structs storing information on which sequences from the original FASTA file were added to each index part
   vector<vector<index_parts_stats> > index_parts_stats_vec;
@@ -1660,27 +623,27 @@ paralleltraversal ( char* inputreads,
   for (uint32_t i = 0; i < myfiles.size()-1; i++) skiplengths.push_back(skiplengths[0]);
     
   // add header lines to SAM output file
-  preprocess_data( myfiles,
-                  argv,
-                  argc,
-                  yes_SQ,
-                  acceptedstrings_sam,
-                  match,
-                  mismatch,
-                  gap_open,
-                  gap_extension,
-                  skiplengths,
-                  num_index_parts,
-                  index_parts_stats_vec,
-                  full_ref,
-                  full_read,
-                  lnwin,
-                  partialwin,
-                  minimal_score,
-                  number_total_read,
-                  gumbel,
-                  numbvs,
-                  numseq);
+  load_index_stats( myfiles,
+                    argv,
+                    argc,
+                    yes_SQ,
+                    acceptedstrings_sam,
+                    match,
+                    mismatch,
+                    gap_open,
+                    gap_extension,
+                    skiplengths,
+                    num_index_parts,
+                    index_parts_stats_vec,
+                    full_ref,
+                    full_read,
+                    lnwin,
+                    partialwin,
+                    minimal_score,
+                    number_total_read,
+                    gumbel,
+                    numbvs,
+                    numseq);
     
   // some info on chosen parameters
   eprintf("  Parameters summary:\n");
